Output beam search time consumption
