--- llm_bench/python/utils/hook_transformers.py
@@ -1,324 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright (C) 2018-2023 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-# flake8: noqa
-import time
-import torch
-import warnings
-import transformers
-import torch.distributed as dist
-import logging as log
-from packaging import version
-from typing import Optional, Tuple, Union, List
-from transformers.generation.stopping_criteria import (
-    StoppingCriteriaList,
-    validate_stopping_criteria,
-)
-from transformers.generation.logits_process import LogitsProcessorList
-from transformers.generation.streamers import BaseStreamer
-from transformers.utils import ModelOutput
-
-
-class GreedySearchDecoderOnlyOutput(ModelOutput):
-    sequences: torch.LongTensor = None
-    scores: Optional[Tuple[torch.FloatTensor]] = None
-    attentions: Optional[Tuple[Tuple[torch.FloatTensor]]] = None
-    hidden_states: Optional[Tuple[Tuple[torch.FloatTensor]]] = None
-
-
-class GreedySearchEncoderDecoderOutput(ModelOutput):
-    sequences: torch.LongTensor = None
-    scores: Optional[Tuple[torch.FloatTensor]] = None
-    encoder_attentions: Optional[Tuple[torch.FloatTensor]] = None
-    encoder_hidden_states: Optional[Tuple[torch.FloatTensor]] = None
-    decoder_attentions: Optional[Tuple[Tuple[torch.FloatTensor]]] = None
-    cross_attentions: Optional[Tuple[Tuple[torch.FloatTensor]]] = None
-    decoder_hidden_states: Optional[Tuple[Tuple[torch.FloatTensor]]] = None
-
-
-GreedySearchOutput = Union[GreedySearchEncoderDecoderOutput, GreedySearchDecoderOnlyOutput]
-
-tm_list = []
-tm_infer_list = []
-
-
-# Copied from https://github.com/huggingface/transformers/blob/6da93f5580e109fad5f7b523cf2b6e8a5bafb623/src/transformers/generation/utils.py#L2282
-# Add the function of collecting latency
-def new_greedy_search(
-    self,
-    input_ids: torch.LongTensor,
-    logits_processor: Optional[LogitsProcessorList] = None,
-    stopping_criteria: Optional[StoppingCriteriaList] = None,
-    max_length: Optional[int] = None,
-    pad_token_id: Optional[int] = None,
-    eos_token_id: Optional[Union[int, List[int]]] = None,
-    output_attentions: Optional[bool] = None,
-    output_hidden_states: Optional[bool] = None,
-    output_scores: Optional[bool] = None,
-    return_dict_in_generate: Optional[bool] = None,
-    synced_gpus: Optional[bool] = False,
-    streamer: Optional["BaseStreamer"] = None,
-    **model_kwargs,
-) -> Union[GreedySearchOutput, torch.LongTensor]:
-    r"""
-    Generates sequences of token ids for models with a language modeling head using **greedy decoding** and can be
-    used for text-decoder, text-to-text, speech-to-text, and vision-to-text models.
-
-    <Tip warning={true}>
-
-    In most cases, you do not need to call [`~generation.GenerationMixin.greedy_search`] directly. Use generate()
-    instead. For an overview of generation strategies and code examples, check the [following
-    guide](../generation_strategies).
-
-    </Tip>
-
-
-    Parameters:
-        input_ids (`torch.LongTensor` of shape `(batch_size, sequence_length)`):
-            The sequence used as a prompt for the generation.
-        logits_processor (`LogitsProcessorList`, *optional*):
-            An instance of [`LogitsProcessorList`]. List of instances of class derived from [`LogitsProcessor`]
-            used to modify the prediction scores of the language modeling head applied at each generation step.
-        stopping_criteria (`StoppingCriteriaList`, *optional*):
-            An instance of [`StoppingCriteriaList`]. List of instances of class derived from [`StoppingCriteria`]
-            used to tell if the generation loop should stop.
-
-        max_length (`int`, *optional*, defaults to 20):
-            **DEPRECATED**. Use `logits_processor` or `stopping_criteria` directly to cap the number of generated
-            tokens. The maximum length of the sequence to be generated.
-        pad_token_id (`int`, *optional*):
-            The id of the *padding* token.
-        eos_token_id (`Union[int, List[int]]`, *optional*):
-            The id of the *end-of-sequence* token. Optionally, use a list to set multiple *end-of-sequence* tokens.
-        output_attentions (`bool`, *optional*, defaults to `False`):
-            Whether or not to return the attentions tensors of all attention layers. See `attentions` under
-            returned tensors for more details.
-        output_hidden_states (`bool`, *optional*, defaults to `False`):
-            Whether or not to return the hidden states of all layers. See `hidden_states` under returned tensors
-            for more details.
-        output_scores (`bool`, *optional*, defaults to `False`):
-            Whether or not to return the prediction scores. See `scores` under returned tensors for more details.
-        return_dict_in_generate (`bool`, *optional*, defaults to `False`):
-            Whether or not to return a [`~utils.ModelOutput`] instead of a plain tuple.
-        synced_gpus (`bool`, *optional*, defaults to `False`):
-            Whether to continue running the while loop until max_length (needed for ZeRO stage 3)
-        streamer (`BaseStreamer`, *optional*):
-            Streamer object that will be used to stream the generated sequences. Generated tokens are passed
-            through `streamer.put(token_ids)` and the streamer is responsible for any further processing.
-        model_kwargs:
-            Additional model specific keyword arguments will be forwarded to the `forward` function of the model.
-            If model is an encoder-decoder model the kwargs should include `encoder_outputs`.
-
-    Return:
-        [`~generation.GreedySearchDecoderOnlyOutput`], [`~generation.GreedySearchEncoderDecoderOutput`] or
-        `torch.LongTensor`: A `torch.LongTensor` containing the generated tokens (default behaviour) or a
-        [`~generation.GreedySearchDecoderOnlyOutput`] if `model.config.is_encoder_decoder=False` and
-        `return_dict_in_generate=True` or a [`~generation.GreedySearchEncoderDecoderOutput`] if
-        `model.config.is_encoder_decoder=True`.
-
-    Examples:
-
-    ```python
-    >>> from transformers import (
-    ...     AutoTokenizer,
-    ...     AutoModelForCausalLM,
-    ...     LogitsProcessorList,
-    ...     MinLengthLogitsProcessor,
-    ...     StoppingCriteriaList,
-    ...     MaxLengthCriteria,
-    ... )
-
-    >>> tokenizer = AutoTokenizer.from_pretrained("gpt2")
-    >>> model = AutoModelForCausalLM.from_pretrained("gpt2")
-
-    >>> # set pad_token_id to eos_token_id because GPT2 does not have a PAD token
-    >>> model.generation_config.pad_token_id = model.generation_config.eos_token_id
-
-    >>> input_prompt = "It might be possible to"
-    >>> input_ids = tokenizer(input_prompt, return_tensors="pt").input_ids
-
-    >>> # instantiate logits processors
-    >>> logits_processor = LogitsProcessorList(
-    ...     [
-    ...         MinLengthLogitsProcessor(10, eos_token_id=model.generation_config.eos_token_id),
-    ...     ]
-    ... )
-    >>> stopping_criteria = StoppingCriteriaList([MaxLengthCriteria(max_length=20)])
-
-    >>> outputs = model.greedy_search(
-    ...     input_ids, logits_processor=logits_processor, stopping_criteria=stopping_criteria
-    ... )
-
-    >>> tokenizer.batch_decode(outputs, skip_special_tokens=True)
-    ["It might be possible to get a better understanding of the nature of the problem, but it's not"]
-    ```"""
-    # init values
-    logits_processor = logits_processor if logits_processor is not None else LogitsProcessorList()
-    stopping_criteria = stopping_criteria if stopping_criteria is not None else StoppingCriteriaList()
-    if max_length is not None:
-        warnings.warn(
-            "`max_length` is deprecated in this function, use" " `stopping_criteria=StoppingCriteriaList([MaxLengthCriteria(max_length=max_length)])` instead.",
-            UserWarning,
-        )
-        stopping_criteria = validate_stopping_criteria(stopping_criteria, max_length)
-    pad_token_id = pad_token_id if pad_token_id is not None else self.generation_config.pad_token_id
-    eos_token_id = eos_token_id if eos_token_id is not None else self.generation_config.eos_token_id
-    if isinstance(eos_token_id, int):
-        eos_token_id = [eos_token_id]
-    eos_token_id_tensor = torch.tensor(eos_token_id).to(input_ids.device) if eos_token_id is not None else None
-    output_scores = output_scores if output_scores is not None else self.generation_config.output_scores
-    output_attentions = output_attentions if output_attentions is not None else self.generation_config.output_attentions
-    output_hidden_states = output_hidden_states if output_hidden_states is not None else self.generation_config.output_hidden_states
-    return_dict_in_generate = return_dict_in_generate if return_dict_in_generate is not None else self.generation_config.return_dict_in_generate
-
-    # init attention / hidden states / scores tuples
-    scores = () if (return_dict_in_generate and output_scores) else None
-    decoder_attentions = () if (return_dict_in_generate and output_attentions) else None
-    cross_attentions = () if (return_dict_in_generate and output_attentions) else None
-    decoder_hidden_states = () if (return_dict_in_generate and output_hidden_states) else None
-
-    # if model is an encoder-decoder, retrieve encoder attention weights and hidden states
-    if return_dict_in_generate and self.config.is_encoder_decoder:
-        encoder_attentions = model_kwargs["encoder_outputs"].get("attentions") if output_attentions else None
-        encoder_hidden_states = model_kwargs["encoder_outputs"].get("hidden_states") if output_hidden_states else None
-
-    # keep track of which sequences are already finished
-    unfinished_sequences = torch.ones(input_ids.shape[0], dtype=torch.long, device=input_ids.device)
-
-    this_peer_finished = False  # used by synced_gpus only
-    while True:
-        tic = time.perf_counter()
-        if synced_gpus:
-            # Under synced_gpus the `forward` call must continue until all gpus complete their sequence.
-            # The following logic allows an early break if all peers finished generating their sequence
-            this_peer_finished_flag = torch.tensor(0.0 if this_peer_finished else 1.0).to(input_ids.device)
-            # send 0.0 if we finished, 1.0 otherwise
-            dist.all_reduce(this_peer_finished_flag, op=dist.ReduceOp.SUM)
-            # did all peers finish? the reduced sum will be 0.0 then
-            if this_peer_finished_flag.item() == 0.0:
-                break
-
-        # prepare model inputs
-        model_inputs = self.prepare_inputs_for_generation(input_ids, **model_kwargs)
-
-        # forward pass to get next token
-        tic_infer = time.perf_counter()
-        outputs = self(
-            **model_inputs,
-            return_dict=True,
-            output_attentions=output_attentions,
-            output_hidden_states=output_hidden_states,
-        )
-        tm_infer_list.append(time.perf_counter() - tic_infer)
-
-        if synced_gpus and this_peer_finished:
-            continue  # don't waste resources running the code we don't need
-
-        next_token_logits = outputs.logits[:, -1, :]
-
-        # pre-process distribution
-        next_tokens_scores = logits_processor(input_ids, next_token_logits)
-
-        # Store scores, attentions and hidden_states when required
-        if return_dict_in_generate:
-            if output_scores:
-                scores += (next_tokens_scores,)
-            if output_attentions:
-                decoder_attentions += (outputs.decoder_attentions,) if self.config.is_encoder_decoder else (outputs.attentions,)
-                if self.config.is_encoder_decoder:
-                    cross_attentions += (outputs.cross_attentions,)
-
-            if output_hidden_states:
-                decoder_hidden_states += (outputs.decoder_hidden_states,) if self.config.is_encoder_decoder else (outputs.hidden_states,)
-
-        # argmax
-        next_tokens = torch.argmax(next_tokens_scores, dim=-1)
-
-        # finished sentences should have their next token be a padding token
-        if eos_token_id is not None:
-            if pad_token_id is None:
-                raise ValueError("If `eos_token_id` is defined, make sure that `pad_token_id` is defined.")
-            next_tokens = next_tokens * unfinished_sequences + pad_token_id * (1 - unfinished_sequences)
-
-        # update generated ids, model inputs, and length for next step
-        input_ids = torch.cat([input_ids, next_tokens[:, None]], dim=-1)
-        if streamer is not None:
-            streamer.put(next_tokens.cpu())
-        model_kwargs = self._update_model_kwargs_for_generation(outputs, model_kwargs, is_encoder_decoder=self.config.is_encoder_decoder)
-
-        # if eos_token was found in one sentence, set sentence to finished
-        if eos_token_id_tensor is not None:
-            unfinished_sequences = unfinished_sequences.mul(next_tokens.tile(eos_token_id_tensor.shape[0], 1).ne(eos_token_id_tensor.unsqueeze(1)).prod(dim=0))
-
-            # stop when each sentence is finished
-            if unfinished_sequences.max() == 0:
-                this_peer_finished = True
-
-        # stop if we exceed the maximum length
-        if stopping_criteria(input_ids, scores):
-            this_peer_finished = True
-        tm_list.append(time.perf_counter() - tic)
-        if this_peer_finished and not synced_gpus:
-            break
-
-    if streamer is not None:
-        streamer.end()
-
-    if return_dict_in_generate:
-        if self.config.is_encoder_decoder:
-            return GreedySearchEncoderDecoderOutput(
-                sequences=input_ids,
-                scores=scores,
-                encoder_attentions=encoder_attentions,
-                encoder_hidden_states=encoder_hidden_states,
-                decoder_attentions=decoder_attentions,
-                cross_attentions=cross_attentions,
-                decoder_hidden_states=decoder_hidden_states,
-            )
-        else:
-            return GreedySearchDecoderOnlyOutput(
-                sequences=input_ids,
-                scores=scores,
-                attentions=decoder_attentions,
-                hidden_states=decoder_hidden_states,
-            )
-    else:
-        return input_ids
-
-
-class BenchHook:
-    def __init__(self):
-        """Clear the time list."""
-        global tm_list
-        tm_list.clear()
-        global tm_infer_list
-        tm_infer_list.clear()
-
-    def clear_time_list(self):
-        """Clear the time list."""
-        global tm_list
-        tm_list.clear()
-
-    def get_time_list(self):
-        """Return the time list."""
-        return tm_list
-
-    def clear_time_infer_list(self):
-        """Clear the infer time list."""
-        global tm_infer_list
-        tm_infer_list.clear()
-
-    def get_time_infer_list(self):
-        """Return the infer time list."""
-        global tm_infer_list
-        return tm_infer_list
-
-    def new_forward(self, model, model_type=None):
-        """Define a new greedy search function."""
-        min_version = version.parse('4.28.0')
-        trans_version = version.parse(transformers.__version__)
-        if trans_version < min_version:
-            log.warning(f'The function of getting latency will not be available with current transformers version:{trans_version}')
-        else:
-            bound_method = new_greedy_search.__get__(model, model.__class__)
-            model.greedy_search = bound_method
